requirements.txt
Django==5.1.5
celery==5.6.2
daphne==4.2.1
django-allauth==65.13.1
django-celery-beat>=2.6
channels>=4.0
channels-redis>=4.0
mysqlclient>=2.2
Pillow>=10.0
redis>=5.0
requests==2.32.5
PyJWT>=2.8
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
